requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
python-dotenv
streamlit
requests
google-generativeai
langchain-ollama
pydantic